Resolve bugs in file

+++ application/views/admin/itemcategory/itemcategoryEdit.php
@@ -3,12 +3,13 @@
     <!-- Content Header (Page header) -->
     <section class="content-header">
         <h1>
-            <i class="fa fa-object-group"></i> <?php echo $this->lang->line('item_category'); ?></h1>
+            <i class="fa fa-object-group"></i> <?php echo $this->lang->line('item_category'); ?>
+        </h1>
     </section>
     <!-- Main content -->
     <section class="content">
         <div class="row">
-            <?php if ($this->rbac->hasPrivilege('item_category', 'can_add') || $this->rbac->hasPrivilege('item_category', 'can_edit')) { ?> 
+            <?php if ($this->rbac->hasPrivilege('item_category', 'can_add') || $this->rbac->hasPrivilege('item_category', 'can_edit')) { ?>
                 <div class="col-md-4">
                     <!-- Horizontal Form -->
                     <div class="box box-primary">
@@ -16,13 +17,24 @@
                             <h3 class="box-title"><?php echo $this->lang->line('edit_item_category'); ?></h3>
                         </div><!-- /.box-header -->
                         <!-- form start -->
-                        <form action="<?php echo site_url("admin/itemcategory/edit/" . $id) ?>"  id="employeeform" name="employeeform" method="post" accept-charset="utf-8">
+                        <form action="<?php echo site_url("admin/itemcategory/edit/" . $id) ?>" id="employeeform" name="employeeform" method="post" accept-charset="utf-8">
                             <div class="box-body">
                                 <?php echo validation_errors(); ?>
                                 <?php echo $this->customlib->getCSRF(); ?>
+                                <div class='form-group'>
+                                    <label for='exampleInputEmail1'><?php echo $this->lang->line('branch'); ?></label><small class='req'> *</small>
+                                    <select id='branch_id' name='branch_id' placeholder='' type='text' class='form-control'>
+                                        <?php foreach ($branch as $key => $value) {  ?>
+                                            <option value='<?php echo $value['id'] ?>' <?php if ($value['id'] == $itemcategory['branch_id']) {
+                                                                                            echo "selected";
+                                                                                        } ?>><?php echo $value['branch_name'] ?></option>
+                                        <?php } ?>
+                                    </select>
+                                    <span class='text-danger'><?php echo form_error('branch'); ?></span>
+                                </div>
                                 <div class="form-group">
                                     <label for="exampleInputEmail1"> <?php echo $this->lang->line('item_category'); ?></label><small class="req"> *</small>
-                                    <input autofocus="" id="itemcategory" name="itemcategory" placeholder="itemcategory" type="text" class="form-control"  value="<?php echo set_value('itemcategory', $itemcategory['item_category']); ?>" />
+                                    <input autofocus="" id="itemcategory" name="itemcategory" placeholder="itemcategory" type="text" class="form-control" value="<?php echo set_value('itemcategory', $itemcategory['item_category']); ?>" />
                                     <span class="text-danger"><?php echo form_error('itemcategory'); ?></span>
                                 </div>
                                 <div class="form-group">
@@ -37,15 +49,16 @@
                             </div>
                         </form>
                     </div>
-                </div><!--/.col (right) -->
+                </div>
+                <!--/.col (right) -->
             <?php } ?>
             <div class="col-md-<?php
-            if ($this->rbac->hasPrivilege('item_category', 'can_add') || $this->rbac->hasPrivilege('item_category', 'can_edit')) {
-                echo "8";
-            } else {
-                echo "12";
-            }
-            ?>">
+                                if ($this->rbac->hasPrivilege('item_category', 'can_add') || $this->rbac->hasPrivilege('item_category', 'can_edit')) {
+                                    echo "8";
+                                } else {
+                                    echo "12";
+                                }
+                                ?>">
                 <!-- general form elements -->
                 <div class="box box-primary" id="exphead">
                     <div class="box-header ptbnull">
@@ -54,56 +67,61 @@
                     <div class="box-body  ">
                         <div class="mailbox-messages">
                             <div class="download_label"><?php echo $this->lang->line('item_category_list'); ?></div>
-                            <div class="table-responsive">  
+                            <div class="table-responsive">
                                 <table class="table table-striped table-bordered table-hover example">
                                     <thead>
                                         <tr>
                                             <th><?php echo $this->lang->line('item_category'); ?></th>
+                                            <th><?php echo $this->lang->line('branch'); ?></th>
                                             <th class="text-right no-print"><?php echo $this->lang->line('action'); ?></th>
                                         </tr>
                                     </thead>
                                     <tbody>
                                         <?php if (empty($categorylist)) {
-                                            ?>
+                                        ?>
 
                                             <?php
                                         } else {
                                             $count = 1;
                                             foreach ($categorylist as $category) {
-                                                ?>
-                                                <tr>                                               
+                                            ?>
+                                                <tr>
                                                     <td class="mailbox-name">
-                                                        <a href="#" data-toggle="popover" class="detail_popover" >
+                                                        <a href="#" data-toggle="popover" class="detail_popover">
                                                             <?php echo $category['item_category'] ?>
                                                         </a>
 
                                                         <div class="fee_detail_popover" style="display: none">
                                                             <?php
                                                             if ($category['description'] == "") {
-                                                                ?>
+                                                            ?>
                                                                 <p class="text text-danger"><?php echo $this->lang->line('no_description'); ?></p>
-                                                                <?php
+                                                            <?php
                                                             } else {
-                                                                ?>
+                                                            ?>
                                                                 <p class="text text-info"><?php echo $category['description']; ?></p>
-                                                                <?php
+                                                            <?php
                                                             }
                                                             ?>
                                                         </div>
                                                     </td>
+                                                    <td>
+                                                        <?php echo $category['branch_name'] ?>
+                                                    </td>
                                                     <td class="mailbox-date pull-right no-print">
-                                                        <?php if ($this->rbac->hasPrivilege('item_category', 'can_edit')) { ?> 
-                                                            <a data-placement="left" href="<?php echo base_url(); ?>admin/itemcategory/edit/<?php echo $category['id'] ?>" class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
+                                                        <?php if ($this->rbac->hasPrivilege('item_category', 'can_edit')) { ?>
+                                                            <a data-placement="left" href="<?php echo base_url(); ?>admin/itemcategory/edit/<?php echo $category['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('edit'); ?>">
                                                                 <i class="fa fa-pencil"></i>
                                                             </a>
-                                                        <?php } if ($this->rbac->hasPrivilege('item_category', 'can_delete')) { ?> 
-                                                            <a data-placement="left" href="<?php echo base_url(); ?>admin/itemcategory/delete/<?php echo $category['id'] ?>"class="btn btn-default btn-xs"  data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
+                                                        <?php }
+                                                        if ($this->rbac->hasPrivilege('item_category', 'can_delete')) { ?>
+                                                            <a data-placement="left" href="<?php echo base_url(); ?>admin/itemcategory/delete/<?php echo $category['id'] ?>" class="btn btn-default btn-xs" data-toggle="tooltip" title="<?php echo $this->lang->line('delete'); ?>" onclick="return confirm('<?php echo $this->lang->line('delete_confirm') ?>');">
                                                                 <i class="fa fa-remove"></i>
                                                             </a>
                                                         <?php } ?>
                                                     </td>
                                                 </tr>
-                                                <?php
+                                        <?php
                                             }
                                             $count++;
                                         }
@@ -111,24 +129,24 @@
 
                                     </tbody>
                                 </table><!-- /.table -->
-                            </div>  
+                            </div>
                         </div><!-- /.mail-box-messages -->
                     </div><!-- /.box-body -->
                 </div>
             </div>
 
 
-        </div>   <!-- /.row -->
+        </div> <!-- /.row -->
     </section><!-- /.content -->
 </div><!-- /.content-wrapper -->
 <script>
-    $(document).ready(function () {
+    $(document).ready(function() {
         $('.detail_popover').popover({
             placement: 'right',
             trigger: 'hover',
             container: 'body',
             html: true,
-            content: function () {
+            content: function() {
                 return $(this).closest('td').find('.fee_detail_popover').html();
             }
         });
